Add common abilities with choice dependencies
New abilities.py module in archivematicaCommon defines abilities of an
Archivematica instance as a whole. An ability can have dependencies. For
example, the Backlog feature is dependent on the Search ability. If
the Search (i.e., Elasticsearch indexing) ability is not enabled, then the
Backlog tab should be unavailable; in addition, the choice to send a transfer to
backlog (either via pre-configuration or via real-time choice in the
GUI) should not be available either.

Other changes:

- The DISABLE_SEARCH_INDEXING boolean configuration option was replaced
  with SEARCH_ENABLED.
- Minor formatting and i18n fixes.