Settings configs support two search switches
For backwards compatibility, the ``disable_search_indexing``
configuration option is restored and is made to co-exist with the new
``search_enabled`` config option. If the former is set, its value
determines the value of the authoritative ``SEARCH_ENABLED`` setting
attribute. If not, then the value of the latter takes effect, if set. If
neither is set, the default value is used.